DASopt.jl

The optimization wrappers

try_many is designed to evaluate a function on many random inputs.

optim_wrap is a wrapper around Optim.

optim_wrap_many uses try_many to call optim_wrap from many random initial points.

To the extent that parameters of optim_wrap_many are inherited from optim_wrap and try_many, they are described in those routines.

DASopt.try_many — Function

x, val = try_many(obj::Function, gen::Function, sense::Symbol; n_tries = Inf, t_lim = Inf, ...

Return the best of many calls of obj on the output of gen, where best is according to sense which should be :Max or :Min. Repeats for n_tries trials or until t_lim seconds pass. One of these must be set.

Arguments

obj::Function: The objective function

gen::Function: A zero-argument function. gen() should generate an input to obj. For example, ()->rand(4).

sense::Symbol: :Min or :Max

n::Integer: An upper bound on the number of inputs to try.

t_lim: Will stop after t_lim seconds.

fn_base: if this is non-empty, then output will be saved to $(fn_base).txt and $(fn_base).jld.

verbosity::Integer: if 0, should be no output. 1, will output just the best so far each time a new one is found. 2, will output everything.

seed::Bool: if negative, nothing happens. Otherwise, seed the ith trial with the ith random seed.

local_rng::Bool: if true, gen() should take a RNG as input, and it will be seeded as suggested by seed. This prevents the gen erator for instances from intefering with any generator the algorithm might use.

verbose::Bool: for backwards compatibility. Used to set verbosity.

par_batch::Integer: if 0, don't parallelize. If > 0, run in batches of this size. Will force seed, unless using localrng. localrng parallelizes the time of generation.

stop_val: will stop as soon as find a value better than this

DASopt.optim_wrap — Function

x, val = optim_wrap(f, x0, mapin;
nrounds=3,
optfunc=NelderMead(),
sense = :Max,
options = Optim.Options(),
n_starts = 0)

Arguments

f is a nonlinear func to optimize

x0 is an initial point. Instead of being a vector, x0 can be a generator, like ()->rand(3)

mapin is a function applied to things like x to put in range.

We run for nrounds, with a mapinat the end of each to polish.

The n_starts option then tells us to start from the best of n_startsruns.

DASopt.optim_wrap_many — Function

x, val = optim_wrap_many(f, x0, mapin; n = Inf, t_lim = Inf,
    nrounds=3,
    optfunc=NelderMead(),
    sense = :Max,
    options = Optim.Options(),
    n_starts = 0,
    par_batch = 0)

val, best = optim_wrap_many(f, x0fun, mapin, nruns;
nrounds, optfunc, maxevals, sense,
file=[])

Call optim_wrap inside try_many. Kwargs come from both. At least one of n or t_lim must be set.

Other optimization heuristics

DASopt.popevolve — Function

opt = popevolve(f, x0, tlim; n = 10, mapin = identity)

Generates and evolves a population of potential solutions to minimize the function f.

n is the number of populations elements

tlim the time limit in seconds

The return, opt, has a couple of fields:

minimizer: the best solution

minimum: the value of it

pop: the population at the end